refactor: move `detectors.geometry` submodule into `detectors` module
