+++ docs/api/pybes3.detectors.md
@@ -14,123 +14,76 @@
 ::: pybes3.detectors.parse_cgem_digi_id
 ---
 
-## Geometry
-::: pybes3.detectors.geometry.get_mdc_wire_position
+## MDC
+::: pybes3.detectors.get_mdc_wire_position
 ---
-::: pybes3.detectors.geometry.get_mdc_gid
+::: pybes3.detectors.get_mdc_gid
 ---
-::: pybes3.detectors.geometry.mdc_gid_to_layer
+::: pybes3.detectors.mdc_gid_to_layer
 ---
-::: pybes3.detectors.geometry.mdc_gid_to_wire
+::: pybes3.detectors.mdc_gid_to_wire
 ---
-::: pybes3.detectors.geometry.mdc_gid_to_is_stereo
+::: pybes3.detectors.mdc_gid_to_is_stereo
 ---
-::: pybes3.detectors.geometry.mdc_layer_to_is_stereo
+::: pybes3.detectors.mdc_layer_to_is_stereo
 ---
-::: pybes3.detectors.geometry.mdc_gid_to_east_x
+::: pybes3.detectors.mdc_gid_to_east_x
 ---
-::: pybes3.detectors.geometry.mdc_gid_to_east_y
+::: pybes3.detectors.mdc_gid_to_east_y
 ---
-::: pybes3.detectors.geometry.mdc_gid_to_east_z
+::: pybes3.detectors.mdc_gid_to_east_z
 ---
-::: pybes3.detectors.geometry.mdc_gid_to_west_x
+::: pybes3.detectors.mdc_gid_to_west_x
 ---
-::: pybes3.detectors.geometry.mdc_gid_to_west_y
+::: pybes3.detectors.mdc_gid_to_west_y
 ---
-::: pybes3.detectors.geometry.mdc_gid_to_west_z
+::: pybes3.detectors.mdc_gid_to_west_z
 ---
-::: pybes3.detectors.geometry.mdc_gid_z_to_x
+::: pybes3.detectors.mdc_gid_z_to_x
 ---
-::: pybes3.detectors.geometry.mdc_gid_z_to_y
----
-::: pybes3.detectors.geometry.get_emc_crystal_position
----
-::: pybes3.detectors.geometry.get_emc_gid
----
-::: pybes3.detectors.geometry.emc_gid_to_center_x
----
-::: pybes3.detectors.geometry.emc_gid_to_center_y
----
-::: pybes3.detectors.geometry.emc_gid_to_center_z
----
-::: pybes3.detectors.geometry.emc_gid_to_front_center_x
----
-::: pybes3.detectors.geometry.emc_gid_to_front_center_y
----
-::: pybes3.detectors.geometry.emc_gid_to_front_center_z
----
-::: pybes3.detectors.geometry.emc_gid_to_part
----
-::: pybes3.detectors.geometry.emc_gid_to_phi
----
-::: pybes3.detectors.geometry.emc_gid_to_point_x
----
-::: pybes3.detectors.geometry.emc_gid_to_point_y
----
-::: pybes3.detectors.geometry.emc_gid_to_point_z
----
-::: pybes3.detectors.geometry.emc_gid_to_theta
+::: pybes3.detectors.mdc_gid_z_to_y
 ---
 
-
-## Digi Identifier
-::: pybes3.detectors.digi_id.check_mdc_id
+## EMC
+::: pybes3.detectors.emc_barrel_h1
 ---
-::: pybes3.detectors.digi_id.mdc_id_to_wire
+::: pybes3.detectors.emc_barrel_h2
 ---
-::: pybes3.detectors.digi_id.mdc_id_to_layer
+::: pybes3.detectors.emc_barrel_h3
 ---
-::: pybes3.detectors.digi_id.mdc_id_to_is_stereo
+::: pybes3.detectors.emc_barrel_l
 ---
-::: pybes3.detectors.digi_id.get_mdc_digi_id
+::: pybes3.detectors.emc_barrel_offset_1
 ---
-::: pybes3.detectors.digi_id.check_tof_id
+::: pybes3.detectors.emc_barrel_offset_2
 ---
-::: pybes3.detectors.digi_id.tof_id_to_part
+::: pybes3.detectors.emc_barrel_r
 ---
-::: pybes3.detectors.digi_id.tof_id_to_layer_or_module
+::: pybes3.detectors.get_emc_crystal_position
 ---
-::: pybes3.detectors.digi_id.tof_id_to_phi_or_strip
+::: pybes3.detectors.get_emc_gid
 ---
-::: pybes3.detectors.digi_id.tof_id_to_end
+::: pybes3.detectors.emc_gid_to_center_x
 ---
-::: pybes3.detectors.digi_id.get_tof_digi_id
+::: pybes3.detectors.emc_gid_to_center_y
 ---
-::: pybes3.detectors.digi_id.check_emc_id
+::: pybes3.detectors.emc_gid_to_center_z
 ---
-::: pybes3.detectors.digi_id.emc_id_to_module
+::: pybes3.detectors.emc_gid_to_front_center_x
 ---
-::: pybes3.detectors.digi_id.emc_id_to_theta
+::: pybes3.detectors.emc_gid_to_front_center_y
 ---
-::: pybes3.detectors.digi_id.emc_id_to_phi
+::: pybes3.detectors.emc_gid_to_front_center_z
 ---
-::: pybes3.detectors.digi_id.get_emc_digi_id
+::: pybes3.detectors.emc_gid_to_part
 ---
-::: pybes3.detectors.digi_id.check_muc_id
+::: pybes3.detectors.emc_gid_to_phi
 ---
-::: pybes3.detectors.digi_id.muc_id_to_part
+::: pybes3.detectors.emc_gid_to_point_x
 ---
-::: pybes3.detectors.digi_id.muc_id_to_segment
+::: pybes3.detectors.emc_gid_to_point_y
 ---
-::: pybes3.detectors.digi_id.muc_id_to_layer
+::: pybes3.detectors.emc_gid_to_point_z
 ---
-::: pybes3.detectors.digi_id.muc_id_to_strip
----
-::: pybes3.detectors.digi_id.muc_id_to_gap
----
-::: pybes3.detectors.digi_id.muc_id_to_strip
----
-::: pybes3.detectors.digi_id.get_muc_digi_id
----
-::: pybes3.detectors.digi_id.check_cgem_id
----
-::: pybes3.detectors.digi_id.cgem_id_to_layer
----
-::: pybes3.detectors.digi_id.cgem_id_to_sheet
----
-::: pybes3.detectors.digi_id.cgem_id_to_strip
----
-::: pybes3.detectors.digi_id.cgem_id_to_is_x_strip
----
-::: pybes3.detectors.digi_id.get_cgem_digi_id
+::: pybes3.detectors.emc_gid_to_theta
 ---
