+++ docs/api/pybes3.digi_id.md
@@ -1,0 +1,3 @@
+# pybes3.digi_id
+
+::: pybes3.digi_id
