+++ docs/api/pybes3.md
@@ -87,6 +87,8 @@
 ---
 ::: pybes3.mdc_layer_to_superlayer
 ---
+
+## digi_id
 ::: pybes3.parse_cgem_digi_id
 ---
 ::: pybes3.parse_emc_digi
