+++ mkdocs.yml
@@ -88,6 +88,7 @@ nav:
   - API Reference:
       - pybes3: api/pybes3.md
       - pybes3.besio: api/pybes3.besio.md
+      - pybes3.digi_id: api/pybes3.digi_id.md
       - pybes3.detectors: api/pybes3.detectors.md
       - pybes3.tracks: api/pybes3.tracks.md
       - pybes3._cache_numba: api/pybes3._cache_numba.md
