+++ src/pybes3/data/emc_geom.npz
