+++ src/pybes3/data/mdc_geom.npz
