+++ src/pybes3/detectors/emc.py
@@ -1,15 +1,16 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Literal
+from typing import Any, Literal
 
 import awkward as ak
 import numba as nb
 import numpy as np
 
-from ...typing import FloatLike, IntLike
+from pybes3 import digi_id
+from pybes3.typing import FloatLike, IntLike
 
-_cur_dir = Path(__file__).resolve().parent
+_geom_data_path = Path(__file__).resolve().parent.parent / "data" / "emc_geom.npz"
 
 ENDCAP_PHI_01 = 64
 ENDCAP_PHI_23 = 80
@@ -49,7 +50,7 @@ def _ensure_loaded():
     global _center_x, _center_y, _center_z
     global _front_center_x, _front_center_y, _front_center_z
 
-    _emc_geom = dict(np.load(_cur_dir / "emc_geom.npz"))
+    _emc_geom = dict(np.load(_geom_data_path))
     _part = _emc_geom["part"]
     _theta = _emc_geom["theta"]
     _phi = _emc_geom["phi"]
@@ -388,3 +389,159 @@ for _fn_name in [
 ]:
     globals()[_fn_name] = _make_lazy(globals()[_fn_name])
 del _fn_name
+
+
+def parse_emc_gid(gid: IntLike, with_pos: bool = True) -> ak.Array | dict[str, Any]:
+    """
+    Parse the gid of EMC crystals. "gid" is the global ID of the crystal, ranges from 0 to 6239.
+    When `gid` is an `ak.Array`, the result is an `ak.Array`, otherwise it is a `dict`.
+
+    Keys of the output:
+
+    - `gid`: Global ID of the crystal.
+    - `part`: Part number, 0 for endcap0, 1 for barrel, 2 for endcap1.
+    - `theta`: Theta number.
+    - `phi`: Phi number.
+
+    Optional keys of the output when `with_pos` is `True`:
+
+    - `front_center_x`: x position of the front center of the crystal.
+    - `front_center_y`: y position of the front center of the crystal.
+    - `front_center_z`: z position of the front center of the crystal.
+    - `center_x`: x position of the center of the crystal.
+    - `center_y`: y position of the center of the crystal.
+    - `center_z`: z position of the center of the crystal.
+
+    !!! info
+        The 8 points of the crystal will not be returned here.
+        If you need the 8 points of the crystal, use `emc_gid_to_point_x`, `emc_gid_to_point_y`
+        and `emc_gid_to_point_z`.
+
+    Parameters:
+        gid: The gid of the crystal.
+        with_pos: Whether to include the position information.
+
+    Returns:
+        The parsed result.
+    """
+    part = emc_gid_to_part(gid)
+    theta = emc_gid_to_theta(gid)
+    phi = emc_gid_to_phi(gid)
+
+    res = {"gid": gid, "part": part, "theta": theta, "phi": phi}
+
+    if with_pos:
+        res["front_center_x"] = emc_gid_to_front_center_x(gid)
+        res["front_center_y"] = emc_gid_to_front_center_y(gid)
+        res["front_center_z"] = emc_gid_to_front_center_z(gid)
+        res["center_x"] = emc_gid_to_center_x(gid)
+        res["center_y"] = emc_gid_to_center_y(gid)
+        res["center_z"] = emc_gid_to_center_z(gid)
+
+    if isinstance(gid, ak.Array):
+        return ak.zip(res)
+    else:
+        return res
+
+
+def parse_emc_digi_id(
+    emc_digi_id: IntLike,
+    with_pos: bool = False,
+) -> ak.Array | dict[str, Any]:
+    """
+    Parse EMC digi ID.
+
+    When `emc_digi_id` is an `ak.Array`, the result is an `ak.Array`, otherwise it is a `dict`.
+
+    Keys of the output:
+
+    - `gid`: Global ID of the crystal.
+    - `part`: Part number, 0 for endcap0, 1 for barrel, 2 for endcap1.
+    - `theta`: Theta number.
+    - `phi`: Phi number.
+
+    Optional keys of the output when `with_pos` is `True`:
+
+    - `front_center_x`: x position of the front center of the crystal.
+    - `front_center_y`: y position of the front center of the crystal.
+    - `front_center_z`: z position of the front center of the crystal.
+    - `center_x`: x position of the center of the crystal.
+    - `center_y`: y position of the center of the crystal.
+    - `center_z`: z position of the center of the crystal.
+
+    !!! info
+        The 8 points of the crystal will not be returned here.
+        If you need the 8 points of the crystal, use `emc_gid_to_point_x`, `emc_gid_to_point_y`
+        and `emc_gid_to_point_z`.
+
+    Parameters:
+        emc_digi_id: The EMC digi ID.
+        with_pos: Whether to include the position information.
+
+    Returns:
+        The parsed EMC digi ID.
+
+    """
+    part = digi_id.emc_id_to_module(emc_digi_id)
+    theta = digi_id.emc_id_to_theta(emc_digi_id)
+    phi = digi_id.emc_id_to_phi(emc_digi_id)
+    gid = get_emc_gid(part, theta, phi)
+    return parse_emc_gid(gid, with_pos)
+
+
+def parse_emc_digi(emc_digi: ak.Record, with_pos: bool = False) -> ak.Record:
+    """
+    Parse EMC raw digi array. The raw digi array should contain [`m_intId`,
+    `m_timeChannel`, `m_chargeChannel`, `m_trackIndex`, `m_measure`] fields.
+
+    Fields of the output:
+
+    - `gid`: Global ID of the crystal.
+    - `part`: Part number, 0 for endcap0, 1 for barrel, 2 for endcap1.
+    - `theta`: Theta number.
+    - `phi`: Phi number.
+    - `charge_channel`: Charge channel.
+    - `time_channel`: Time channel.
+    - `track_index`: Track index.
+    - `measure`: Measure value.
+    - `digi_id`: Raw digi ID.
+
+    Optional fields of the output when `with_pos` is `True`:
+
+    - `front_center_x`: x position of the front center of the crystal.
+    - `front_center_y`: y position of the front center of the crystal.
+    - `front_center_z`: z position of the front center of the crystal.
+    - `center_x`: x position of the center of the crystal.
+    - `center_y`: y position of the center of the crystal.
+    - `center_z`: z position of the center of the crystal.
+
+    Parameters:
+        emc_digi: The EMC raw digi array.
+        with_pos: Whether to include the position information.
+
+    Returns:
+        The parsed EMC digi array.
+    """
+    gid = parse_emc_digi_id(emc_digi["m_intId"], with_pos=with_pos)
+
+    res = {
+        "gid": gid["gid"],
+        "part": gid["part"],
+        "theta": gid["theta"],
+        "phi": gid["phi"],
+        "charge_channel": emc_digi["m_chargeChannel"],
+        "time_channel": emc_digi["m_timeChannel"],
+        "track_index": emc_digi["m_trackIndex"],
+        "measure": emc_digi["m_measure"],
+        "digi_id": emc_digi["m_intId"],
+    }
+
+    if with_pos:
+        res["front_center_x"] = gid["front_center_x"]
+        res["front_center_y"] = gid["front_center_y"]
+        res["front_center_z"] = gid["front_center_z"]
+        res["center_x"] = gid["center_x"]
+        res["center_y"] = gid["center_y"]
+        res["center_z"] = gid["center_z"]
+
+    return ak.zip(res)
