--- src/pybes3/detectors/geometry/__init__.py
@@ -1,87 +1,0 @@
-from __future__ import annotations
-
-from .emc import (
-    emc_barrel_h1,
-    emc_barrel_h2,
-    emc_barrel_h3,
-    emc_barrel_l,
-    emc_barrel_offset_1,
-    emc_barrel_offset_2,
-    emc_barrel_r,
-    emc_gid_to_center_x,
-    emc_gid_to_center_y,
-    emc_gid_to_center_z,
-    emc_gid_to_front_center_x,
-    emc_gid_to_front_center_y,
-    emc_gid_to_front_center_z,
-    emc_gid_to_part,
-    emc_gid_to_phi,
-    emc_gid_to_point_x,
-    emc_gid_to_point_y,
-    emc_gid_to_point_z,
-    emc_gid_to_theta,
-    get_emc_crystal_position,
-    get_emc_gid,
-)
-from .mdc import (
-    get_mdc_gid,
-    get_mdc_wire_position,
-    mdc_gid_to_east_x,
-    mdc_gid_to_east_y,
-    mdc_gid_to_east_z,
-    mdc_gid_to_is_stereo,
-    mdc_gid_to_layer,
-    mdc_gid_to_stereo,
-    mdc_gid_to_superlayer,
-    mdc_gid_to_west_x,
-    mdc_gid_to_west_y,
-    mdc_gid_to_west_z,
-    mdc_gid_to_wire,
-    mdc_gid_z_to_x,
-    mdc_gid_z_to_y,
-    mdc_layer_to_is_stereo,
-    mdc_layer_to_superlayer,
-)
-
-__all__ = [
-    # EMC
-    "emc_barrel_h1",
-    "emc_barrel_h2",
-    "emc_barrel_h3",
-    "emc_barrel_l",
-    "emc_barrel_offset_1",
-    "emc_barrel_offset_2",
-    "emc_barrel_r",
-    "emc_gid_to_center_x",
-    "emc_gid_to_center_y",
-    "emc_gid_to_center_z",
-    "emc_gid_to_front_center_x",
-    "emc_gid_to_front_center_y",
-    "emc_gid_to_front_center_z",
-    "emc_gid_to_part",
-    "emc_gid_to_phi",
-    "emc_gid_to_point_x",
-    "emc_gid_to_point_y",
-    "emc_gid_to_point_z",
-    "emc_gid_to_theta",
-    "get_emc_crystal_position",
-    "get_emc_gid",
-    # MDC
-    "get_mdc_gid",
-    "get_mdc_wire_position",
-    "mdc_gid_to_east_x",
-    "mdc_gid_to_east_y",
-    "mdc_gid_to_east_z",
-    "mdc_gid_to_is_stereo",
-    "mdc_gid_to_layer",
-    "mdc_gid_to_stereo",
-    "mdc_gid_to_superlayer",
-    "mdc_gid_to_west_x",
-    "mdc_gid_to_west_y",
-    "mdc_gid_to_west_z",
-    "mdc_gid_to_wire",
-    "mdc_gid_z_to_x",
-    "mdc_gid_z_to_y",
-    "mdc_layer_to_is_stereo",
-    "mdc_layer_to_superlayer",
-]
